requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
cloudinary==1.26.0
Django==3.2.6
django-crispy-forms==1.12.0
django-gravatar2==1.4.4
django-storages==1.11.1
django-widget-tweaks==1.4.8
idna==3.2
lxml==4.6.3
oauthlib==3.1.1
Pillow==8.3.1
PySocks==1.7.1
python-decouple==3.4
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tweepy==3.10.0
urllib3==1.26.6
